requirements.txt
tweepy
pymongo
vaderSentiment
telethon
python-dotenv
apscheduler
openai
